+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Updated
+
+- Added asyncio.sleep to unblock main thread when polling
+
 ## [0.7.1] - 2022-09-13
 
 ### Fixed